Encja Cart Implementacja

+++ src/main/java/com/kodilla/ecommercee/repository/UserDao.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserDao extends CrudRepository<User, Integer> {
+public interface UserDao extends CrudRepository<User, Long> {
     @Override
     User save(User Entity);
 }
